# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a auth server",
        "title": "LoginHub API",
        "contact": {},
        "version": "v1"
    },
    "basePath": "/api/v1",
    "paths": {
        "/passport/device": {
            "get": {
                "description": "获取用户登录的设备列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Passport"
                ],
                "summary": "用户登录设备",
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.RespBody"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/passport.LoginDevice"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数有误",
                        "schema": {
                            "$ref": "#/definitions/handler.RespBody"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/handler.RespBody"
                        }
                    }
                }
            }
        },
        "/passport/device/{id}/kick": {
            "post": {
                "description": "用户注销设备",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Passport"
                ],
                "summary": "用户注销设备",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "设备ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/handler.RespBody"
                        }
                    },
                    "400": {
                        "description": "参数有误",
                        "schema": {
                            "$ref": "#/definitions/handler.RespBody"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/handler.RespBody"
                        }
                    }
                }
            }
        },
        "/passport/login": {
            "post": {
                "description": "根据类型用户登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Passport"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "用户登录请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/passport.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.RespBody"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/passport.LoginResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数有误",
                        "schema": {
                            "$ref": "#/definitions/handler.RespBody"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/handler.RespBody"
                        }
                    }
                }
            }
        },
        "/passport/logout": {
            "post": {
                "description": "用户注销",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Passport"
                ],
                "summary": "用户注销",
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/handler.RespBody"
                        }
                    },
                    "400": {
                        "description": "参数有误",
                        "schema": {
                            "$ref": "#/definitions/handler.RespBody"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/handler.RespBody"
                        }
                    }
                }
            }
        },
        "/passport/mail/send": {
            "post": {
                "description": "根据类型发送电子邮箱验证码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Passport"
                ],
                "summary": "发送电子邮箱验证码",
                "parameters": [
                    {
                        "description": "发送邮箱验证码请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/passport.EmailSendRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "$ref": "#/definitions/handler.RespBody"
                        }
                    },
                    "400": {
                        "description": "参数有误",
                        "schema": {
                            "$ref": "#/definitions/handler.RespBody"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/handler.RespBody"
                        }
                    }
                }
            }
        },
...
        "/passport/register": {
            "post": {
                "description": "根据类型用户注册",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Passport"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "用户注册请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/passport.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.RespBody"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/passport.RegisterResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数有误",
                        "schema": {
                            "$ref": "#/definitions/handler.RespBody"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/handler.RespBody"
                        }
                    }
                }
            }
        },
        "/passport/session/refresh": {
            "post": {
                "description": "刷新会话，返回新的refresh_token并设置cookie",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Passport"
                ],
                "summary": "用户刷新会话",
                "parameters": [
                    {
                        "description": "刷新会话请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/passport.SessionRefreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "请求成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.RespBody"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/passport.SessionRefreshResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "参数有误",
                        "schema": {
                            "$ref": "#/definitions/handler.RespBody"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/handler.RespBody"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.RespBody": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "http code",
                    "type": "integer"
                },
                "data": {
                    "description": "response data"
                },
                "msg": {
                    "description": "response message",
                    "type": "string"
                },
                "reason": {
                    "description": "reason key",
                    "type": "string"
                }
            }
        },
...
        "passport.Device": {
            "type": "object",
            "properties": {
                "browser": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "os": {
                    "type": "string"
                }
            }
        },
        "passport.EmailSendRequest": {
            "type": "object",
            "required": [
                "email",
                "type"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                }
            }
        },
...
        "passport.LoginDevice": {
            "type": "object",
            "properties": {
                "devices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/passport.Device"
                    }
                }
            }
        },
        "passport.LoginRequest": {
            "type": "object",
            "required": [
                "login_type"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "login_type": {
                    "type": "integer",
                    "minimum": 1
                },
                "password": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "provider": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
        "passport.LoginResponse": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/passport.User"
                }
            }
        },
        "passport.RegisterRequest": {
            "type": "object",
            "required": [
                "capcha",
                "password",
                "password_confirm",
                "register_type",
                "user_name"
            ],
            "properties": {
                "capcha": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "password_confirm": {
                    "type": "string"
                },
                "register_type": {
                    "type": "integer",
                    "minimum": 1
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "passport.RegisterResponse": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/passport.User"
                }
            }
        },
        "passport.SessionRefreshRequest": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "passport.SessionRefreshResponse": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
...
        "passport.User": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "uid": {
                    "type": "integer"
                }
            }
        }
    }
}